fix:헤더바 componenet  분리
