+++ src/pages/main/OrderInProgress.jsx
@@ -5,30 +5,15 @@ import OrderProgressCard from "../../components/order/OrderProgressCard";
 import Button from "../../components/common/button/Button";
 import {
   PageContainer,
-  Header,
-  Brand,
-  BrandLogo,
-  HamburgerButton,
   Content,
   OpenStatusSection,
   OpenStatusText,
   SectionTitleWrapper,
   SectionTitle,
   StatusButtons,
-  Backdrop,
-  Drawer,
-  DrawerHeader,
-  ProfileAvatar,
-  ProfileInfo,
-  Nickname,
-  LogoutButton,
-  DrawerList,
-  DrawerItem,
   OrderList,
 } from "./OrderInProgress.styles";
-import greenSquirrelIcon from "../../assets/icons/greensquirrel.png";
-import menuIcon from "../../assets/icons/menu.png";
-import starsquirrelIcon from "../../assets/icons/starsquirrel.png";
+import HeaderSeller from "../../components/common/header/HeaderSeller";
 
 export default function OrderInProgress() {
   const navigate = useNavigate();
@@ -62,17 +47,7 @@ export default function OrderInProgress() {
 
   return (
     <PageContainer>
-      <Header>
-        <Brand>
-          <BrandLogo src={greenSquirrelIcon} alt="심봤다" />
-        </Brand>
-        <HamburgerButton
-          aria-label="메뉴 열기"
-          onClick={() => setDrawerOpen((v) => !v)}
-        >
-          <img src={menuIcon} alt="메뉴" width={24} height={24} />
-        </HamburgerButton>
-      </Header>
+      <HeaderSeller userInfo={userInfo} onLogout={handleLogout} />
 
       <Content>
         <OpenStatusSection>
@@ -136,25 +111,6 @@ export default function OrderInProgress() {
           </OrderProgressCard>
         </OrderList>
       </Content>
-
-      <Backdrop $open={drawerOpen} onClick={() => setDrawerOpen(false)} />
-      <Drawer $open={drawerOpen} ref={drawerRef} aria-hidden={!drawerOpen}>
-        <DrawerHeader>
-          <ProfileAvatar>
-            <img src={starsquirrelIcon} alt="프로필" width={28} height={28} />
-          </ProfileAvatar>
-          <ProfileInfo>
-            <Nickname>{userInfo?.storeName || "로딩 중..."}님</Nickname>
-            <LogoutButton onClick={handleLogout}>로그아웃</LogoutButton>
-          </ProfileInfo>
-        </DrawerHeader>
-        <DrawerList>
-          <DrawerItem>주문 현황</DrawerItem>
-          <DrawerItem onClick={() => navigate("/product-register")}>
-            상품 등록
-          </DrawerItem>
-        </DrawerList>
-      </Drawer>
     </PageContainer>
   );
 }
