+++ src/pages/main/ProductRegister.jsx
@@ -4,25 +4,12 @@ import { getSellerMe, logout } from "../../api/auth";
 import Button from "../../components/common/button/Button";
 import {
   PageContainer,
-  Header,
-  Brand,
-  BrandLogo,
-  HamburgerButton,
   Content,
   OpenStatusSection,
   OpenStatusText,
   SectionTitle,
   SectionTitleWrapper,
   EmptyMessage,
-  Backdrop,
-  Drawer,
-  DrawerHeader,
-  ProfileAvatar,
-  ProfileInfo,
-  Nickname,
-  LogoutButton,
-  DrawerList,
-  DrawerItem,
   FixedBottomButton,
   OrderModal,
   ModalContent,
@@ -51,9 +38,7 @@ import {
   DiscountBox,
 } from "./ProductRegister.styles";
 
-import menuIcon from "../../assets/icons/menu.png";
-import greenSquirrelIcon from "../../assets/icons/greensquirrel.png";
-import starsquirrelIcon from "../../assets/icons/starsquirrel.png";
+import HeaderSeller from "../../components/common/header/HeaderSeller";
 import closeIcon from "../../assets/icons/x.png";
 import ProductCard from "../../components/product/ProductCard";
 
@@ -214,18 +199,7 @@ function ProductRegister() {
 
   return (
     <PageContainer>
-      <Header>
-        <Brand>
-          <BrandLogo src={greenSquirrelIcon} alt="심봤다" />
-        </Brand>
-        <HamburgerButton
-          aria-label="메뉴 열기"
-          onClick={() => setDrawerOpen((v) => !v)}
-        >
-          <img src={menuIcon} alt="메뉴" width={24} height={24} />
-        </HamburgerButton>
-      </Header>
-
+      <HeaderSeller userInfo={userInfo} onLogout={handleLogout} />
       <Content>
         <OpenStatusSection>
           <Button
@@ -273,25 +247,6 @@ function ProductRegister() {
           </Button>
         </FixedBottomButton>
       </Content>
-
-      <Backdrop $open={drawerOpen} onClick={() => setDrawerOpen(false)} />
-      <Drawer $open={drawerOpen} ref={drawerRef} aria-hidden={!drawerOpen}>
-        <DrawerHeader>
-          <ProfileAvatar>
-            <img src={starsquirrelIcon} alt="프로필" width={28} height={28} />
-          </ProfileAvatar>
-          <ProfileInfo>
-            <Nickname>{userInfo?.storeName || "로딩 중..."}님</Nickname>
-            <LogoutButton onClick={handleLogout}>로그아웃</LogoutButton>
-          </ProfileInfo>
-        </DrawerHeader>
-        <DrawerList>
-          <DrawerItem onClick={() => navigate("/mainpage-seller")}>
-            주문 현황
-          </DrawerItem>
-          <DrawerItem>상품 등록</DrawerItem>
-        </DrawerList>
-      </Drawer>
 
       {sheetOpen && (
         <>
